TG changes all new tokens

+++ src/main/java/io/blockchainetl/common/utils/FiatPrices.java
@@ -14,11 +14,11 @@ public class FiatPrices {
     private static final String CURRENCYLAYER_API_KEY = "3395daa1e8a1d505f1ba01f8e4edcbd6";
     private static final Logger LOG =
             LoggerFactory.getLogger(FiatPrices.class);
-    public static LoadingCache<String, Long> coinPrices = CacheBuilder.newBuilder()
+    public static LoadingCache<String, Float> coinPrices = CacheBuilder.newBuilder()
             .expireAfterWrite(24, TimeUnit.HOURS)
             .build(
-                    new CacheLoader<String, Long>() {
-                        public Long load(String key) throws Exception {
+                    new CacheLoader<String, Float>() {
+                        public Float load(String key) throws Exception {
                             return FiatPrices.makeRequest(key);
                         }
                     });
@@ -34,7 +34,7 @@ public class FiatPrices {
         return baseUri;
     }
 
-    public static Long makeRequest(String currencyCode) throws Exception {
+    public static Float makeRequest(String currencyCode) throws Exception {
         try {
             String url = null;
             if (currencyCode.equals("XSGD")) {
@@ -46,13 +46,13 @@ public class FiatPrices {
                 String response = Utils.makeGetRequest(url);
                 JSONObject jsonResponse = new JSONObject(response);
                 LOG.info(jsonResponse.toString());
-                Long price = 1 / jsonResponse.getJSONObject("quotes").getLong("USD" + fiatCurrencyCode);
+                float price = 1 / jsonResponse.getJSONObject("quotes").getLong("USD" + fiatCurrencyCode);
                 return price;
             }
         } catch (Exception e) {
             LOG.info(e.getMessage());
             throw e;
         }
-        return 0L;
+        return 0.0f;
     }
 }
